Changed test folder structure
