+++ src/__test__/components/BoardOptionsMenu/ColourOptions/ColourOptions.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import ColorOptions from "../../../components/BoardOptionsMenu/ColourOptions/ColourOptions";
+import ColorOptions from "../../../../components/BoardOptionsMenu/ColourOptions/ColourOptions";
 
 describe("ColourOptions", () => {
   it("should render successfully", () => {
