+++ src/__test__/components/BoardOptionsMenu/ImageUploadOptions/ImageUploadOptions.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import ImageUploadOptions from "../../../components/BoardOptionsMenu/ImageUploadOptions/ImageUploadOptions";
+import ImageUploadOptions from "../../../../components/BoardOptionsMenu/ImageUploadOptions/ImageUploadOptions";
 
 describe("ImageUploadOptions", () => {
   it("should render successfully", () => {
